feat(builtin, core, util): allow BasicNode to run separately
Move the `util` dir into core, the add-ons in other dir should import utilities from
<rootDir>/index, because some tool func may change their behaviors in different runtime environment.
Change the second arg of [Element].run to `options`, which is an object containing the original
`caller` prop and allow more options. An new optional prop added in BasicNode system is `runStack`,
which is an stack describing the runing of this time. More information about `runStack` will be
offered in the following blog.